+++ queso/io/binary_buffer_writer.hpp
@@ -17,6 +17,7 @@
 //// STL includes
 #include <fstream>
 #include <vector>
+#include <cstring>
 
 //// Project includes
 #include "queso/includes/define.hpp"
@@ -44,7 +45,7 @@ public:
     /// @param rOut Filestream.
     /// @param Endian Options: {big, little}.
     /// @param BufferSize (Default: 1MB).
-    BinaryBufferWriter(std::ofstream& rOut, EndianType Endian, IndexType BufferSize = (int)1 << 20) : // default 1MB
+    BinaryBufferWriter(std::ofstream& rOut, EndianType Endian, IndexType BufferSize = 1ul << 20) : // default 1MB
           mOut(rOut),
           mEndian(Endian),
           mBufferSize(BufferSize),
@@ -103,7 +104,7 @@ public:
     /// @brief Flushes the buffer to the filestream.
     void Flush() {
         if (mBufferPos > 0) {
-            mOut.write(mBuffer.data(), mBufferPos);
+            mOut.write(mBuffer.data(), static_cast<long int>(mBufferPos));
             QuESo_ERROR_IF(!mOut) << "Write failed during Flush." << std::endl;
             mBufferPos = 0; // Reset buffer.
         }
@@ -122,7 +123,7 @@ private:
         if(Size > mBufferSize) {
             Flush();
             // Write large data directly to stream (no buffering).
-            mOut.write(pData, Size);
+            mOut.write(pData, static_cast<long int>(Size));
             QuESo_ERROR_IF(!mOut) << "Write failed during Flush." << std::endl;
             return;
         }
